pyproject.toml
[project]
name = "simple-nl2sql"
version = "0.3.2"
description = "A simple implementation of NL2SQL."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "deprecated>=1.2.18",
    "httpx[socks]>=0.28.1",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.21",
    "openai-agents>=0.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pymilvus>=2.5.11",
    "pymysql>=1.1.1",
    "records>=0.6.0",
    "sqlalchemy>=2.0.41",
]

[project.urls]
Homepage = "https://github.com/secretpot/nl2sql"
Repository = "https://github.com/secretpot/nl2sql"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"nl2sql.tools" = ["resources/**/*"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
